Create Path Analysis
This code is basically to perform SEM and Path analysis on the factors affecting Nigerian Students. 

While that is the aim of the research, it is also useful in business analytics to compare the effect on business decisions on the overall customer satisfaction. 

It helps to understand the magnitude of causation and the direction of causation as well.