Crosstab Result Extraction
the lines of code in the R files was used to extract the valuable information in the exported results from spss to into a table that can be used for presentation purposes. It still requires some manual work on the results itself, but it saves a lot of time from manually inputting the figures to fit the presentation table format.